+++ mayan/apps/ocr/backends/tesseract.py
@@ -2,6 +2,8 @@ from __future__ import unicode_literals
 
 import logging
 
+import sh
+
 from PIL import Image
 import pytesseract
 
@@ -13,6 +15,23 @@ logger = logging.getLogger(__name__)
 
 
 class Tesseract(OCRBackendBase):
+    def __init__(self, *args, **kwargs):
+        super(Tesseract, self).__init__(*args, **kwargs)
+        try:
+            self.binary = sh.Command(setting_tesseract_path.value)
+        except sh.CommandNotFound:
+            self.binary = None
+
+    def get_languages(self):
+        if self.binary:
+            result = self.binary(list_langs=True)
+
+            return [
+                language for language in result.stderr.split('\n') if language
+            ]
+        else:
+            return ()
+
     def execute(self, *args, **kwargs):
         """
         Execute the command line binary of tesseract
@@ -29,7 +48,16 @@ class Tesseract(OCRBackendBase):
             # re-run it with no language parameter
         except Exception as exception:
             error_message = 'Exception calling pytesseract with language option: {}; {}'.format(self.language, exception)
+
+            if self.binary:
+                if self.language not in self.get_languages():
+                    error_message = '{}\nThe requested Tesseract language file for "{}" is not available and needs to be installed.\nIf using Debian or Ubuntu run: apt-get install tesseract-ocr-{}'.format(error_message, self.language, self.language)
+
             logger.error(error_message)
             raise OCRError(error_message)
 
         return result
+
+
+
+